Fix data type issues in __getNyble
In Python 3 the unpack routine fails as it expects bytes to work with but the when a single byte is taken from an array it is an integer so a type mismatch occurs. Since the bit math to get the nibble expects an  integer I changed the code to just take the returned byte.