Added arcsight mode to accompany utf8, hex, and ascii
Added an arcsight mode for the payload field. CEF expects utf8, but the utf8 option includes normal binary bytes that make the payload hard to view. As the payload will be going to a field rather than the payload viewer, the binary bytes do not offer any additional information. The arcsight mode only sends printable characters for ease of human reading in a field.